// types/Bank.ts
export declare namespace Bank {
  export type AccountStruct = {
    owner: PromiseOrValue<string>;
    accountName: PromiseOrValue<string>;
    description: PromiseOrValue<string>;
    amount: PromiseOrValue<BigNumberish>;
    lockTime: PromiseOrValue<BigNumberish>;
    accountId: PromiseOrValue<BigNumberish>;
  };

  export type AccountStructOutput = [
    string,
    string,
    string,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    owner: string;
    accountName: string;
    description: string;
    amount: BigNumber;
    lockTime: BigNumber;
    accountId: BigNumber;
  };
}